送你做宝贝

+++ app/internal/types/types.go
@@ -46,7 +46,7 @@ type CategoryRespSon struct {
 }
 
 type CommonReponse struct {
-	Code int8        `json:"code"`
+	Code int64       `json:"code"`
 	Msg  string      `json:"msg"`
 	Data interface{} `json:"data"`
 }
